+++ CHANGELOG.md
@@ -3,3 +3,9 @@
 ## V0
 
 ### 0.4
+
+#### Feature
+
+- Line height
+- Kerning
+- Vertical Align